fix(code_challenge): refactoring generateCodeVerifier and generateCodeChallenge inside Utils class
